Created Fragment base class and extended via Sentence.
